+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 4.13.0
 
+* feat: add static code diagnostics `prefer-correct-edge-insets-constructor-rule`.
 * feat: add [Checkstyle](https://dartcodemetrics.dev/docs/cli/analyze#checkstyle) format reporter.
 * feat: add [prefer-immediate-return](https://dartcodemetrics.dev/docs/rules/common/prefer-immediate-return) rule